AIGuard: flatten Configuration module into single file
Inline `AIGuard::Configuration::Settings` into
`lib/datadog/ai_guard/configuration.rb` and remove the now-empty
`configuration/settings.rb`. Mirror the RBS layout in `sig/`. The
`configuration/ext.rb` companion file is unchanged; the require
moves up to `configuration.rb`.

Matches the single-file pattern already used by
`lib/datadog/open_feature/configuration.rb`. No behavior change.
