--- sig/datadog/ai_guard/configuration/settings.rbs
@@ -1,11 +1,0 @@
-module Datadog
-  module AIGuard
-    module Configuration
-      module Settings
-        def self.extended: (::Class | ::Module base) -> void
-
-        def self.add_settings!: (untyped base) -> void
-      end
-    end
-  end
-end
